+++ packages/solid/src/segment-group/segment-group.anatomy.ts
@@ -1,0 +1,4 @@
+import { anatomy } from '@zag-js/radio-group'
+
+export const segmentGroupAnatomy = anatomy.rename('segment-group')
+export const parts = segmentGroupAnatomy.build()